Update ostree HEAD and workaround otmp use.
The ostree recipe is updated to point to the latest from the master
branch.

The ostree-native build has an issue where on some configurations (ex:
SUSE 42.1) it gets confused about otmpfile capability, tries to use it
and terminates with a "file not found" error.
For now this is worked around by disabling the use of the otmpfile
feature.
Even the disabling has problems, so, instead of using

--disable-otmpfile = yes   as instructed by the help, it needs
--enable-otmpfile = no

Code from Ismo Puustinen <ismo.puustinen@intel.com>
Signed-off-by: Igor Stoppa <igor.stoppa@intel.com>