[Workaround] Disable QA test for ldflags
lshw seems to not pick up from the environment the value
of LDFLAGS set by bitbake.

The "--hash-style=gnu" option that bitbake would set,is ignored
inside the Makefile and, later on, during the QA phase, this is
detected and causes a QA error.

The current workaround is to disable the test for ldflags.

Signed-off-by: Igor Stoppa <igor.stoppa@intel.com>