ostree_git.bb: add/use GIR-disabling patch, fix build options.
Added a patch to forcibly disable gobject introspection data
generation. It is not needed for more recent versions of
ostree and can be dropped in the future.

Changed config options to explicitly disable features (libarchive)
we do not explicitly depend on. This prevents them from being
picked up accidentially (and QA checks later failing) if/when some
other package pulls them in.

Changed the inverted config options to the more straightforward
--disable-* from --enable-*=no. Seems to work correctly.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>